backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
django-unfold>=0.40,<1.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Azure Speech
azure-cognitiveservices-speech>=1.34,<2.0

# Production
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0

# Development
python-dotenv>=1.0,<2.0
ruff>=0.4,<1.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
